# docs/swagger.yaml
definitions:
  dto.InvestmentDTO:
    properties:
      amount:
        type: number
    required:
    - amount
    type: object
...
  response.ApiResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is my Go Inv API.
  title: My Inv API
  version: "1.0"
paths:
...
  /admin/deposit:
    get:
      description: Fetch all users deposits
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Fetch deposits
      tags:
      - Admin Dashboard
  /admin/deposit/{id}/approve:
    put:
      consumes:
      - application/json
      description: Approve user deposits
      parameters:
      - description: Deposit ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Approve deposit
      tags:
      - Admin Action
  /admin/deposit/{id}/decline:
    put:
      consumes:
      - application/json
      description: Decline user deposits
      parameters:
      - description: id
        example: 550e8400-e29b-41d4-a716-446655440000
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Decline deposit
      tags:
      - Admin Action
  /admin/deposit/pending:
    get:
      description: Fetch all pending users deposits
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Fetch pending deposits
      tags:
      - Admin Dashboard
  /admin/investment/active:
    get:
      description: Fetch Active investment
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Fetch Active investment
      tags:
      - Admin Dashboard
  /admin/kyc:
    get:
      description: Fetch all users kyc data
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Fetch kyc
      tags:
      - Admin Dashboard
  /admin/kyc/{id}/approve:
    put:
      consumes:
      - application/json
      description: Approve user kyc
      parameters:
      - description: kyc ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Approve kyc
      tags:
      - Admin Action
  /admin/kyc/{id}/decline:
    put:
      consumes:
      - application/json
      description: Decline user kyc
      parameters:
      - description: kyc ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Decline kyc
      tags:
      - Admin Action
  /admin/kyc/pending:
    get:
      description: Fetch all pending users kyc data
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Fetch pending kyc
      tags:
      - Admin Dashboard
...
  /user/deposits:
    get:
      description: Returns deposit page data for the logged-in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: User deposit page data
      tags:
      - User Dashboard
    post:
      consumes:
      - multipart/form-data
      description: Create a new deposit for a user
      parameters:
      - description: Deposit Amount
        in: formData
        name: amount
        required: true
        type: number
      - description: Deposit Proof Image
        in: formData
        name: deposit_proof
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: New deposit
      tags:
      - User Action
  /user/investment/{id}:
    post:
      consumes:
      - application/json
      description: User subscribing to an investment plan
      parameters:
      - description: Investment Plan ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: join plan data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.InvestmentDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Subscribe to an investment plan
      tags:
      - User Action
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"